refactor: move three.js file
