+++ frontend/src/game/preset.js
@@ -1,3 +1,3 @@
-import * as THREE from "../three.js";
+import * as THREE from "../threejs/three.js";
 
 export const zaxis = new THREE.Vector3(0, 0, 1);
